feat: add PopupReplace UI

+++ src/components/brandkit/tabs/BrandKitTabLogo.vue
@@ -1,67 +1,32 @@
 <template lang="pug">
-transition-group(class="brand-kit-tab-logo" name="logo-list" tag="div")
-  template(v-for="logo in renderedLogos" :key="logo")
-    div(v-if="logo === 'add'"
-      class="brand-kit-tab-logo__item add pointer relative"
-      key="add"
-      @click="handleUploadLogo")
-      span(class="primary") {{ $t('NN0411') }}
-      span(class="secondary" v-html="$t('NN0412')")
-      svg-icon(class="hover"
-              iconName="plus-origin"
-              iconWidth="16px"
-              iconColor="gray-2")
-    div(v-else-if="logo === 'loading'"
-        class="brand-kit-tab-logo-loading no-trans"
-        key="loading")
-      svg-icon(iconName="loading"
-              iconWidth="50px"
-              iconColor="gray-3")
-    observer-sentinel(v-else-if="logo === 'sentinel'"
-                      class="no-trans"
-                      key="sentinel"
-                      :target="$route.name === 'Editor' ? '.popup-brand-settings__window' : undefined"
-                      @callback="handleLoadMore")
-    div(v-else-if="(typeof logo !== 'string')"
-      class="brand-kit-tab-logo__item relative"
-      :class="{hovered: checkMenuOpen(logo)}"
-      :key="logo.id.replace('new_', '')")
-      svg-icon(v-if="checkUploading(logo)" iconName="loading" iconWidth="24px" iconColor="gray-3")
-      img(v-else :src="getUrl(logo)" class="brand-kit-tab-logo__item__img")
-      div(v-if="!checkUploading(logo)" class="brand-kit-tab-logo__item__more pointer"
-        @click="handleOpenMenu(logo)")
-        div(class="brand-kit-tab-logo__item__more-container relative")
-          svg-icon(iconName="more_vertical"
-                  iconWidth="24px"
-                  iconColor="gray-2")
-          div(v-if="checkMenuOpen(logo)"
-            class="brand-kit-tab-logo__item__menu"
-            v-click-outside="() => { menuOpenLogoId = '' }")
-            div(class="brand-kit-tab-logo__item__menu__name")
-              span {{ logo.name }}
-            div(class="brand-kit-tab-logo__item__menu__hr")
-            div(class="brand-kit-tab-logo__item__menu__row pointer"
-              @click="handleDownload(logo)")
-              svg-icon(iconName="download"
-                      iconWidth="24px"
-                      iconColor="gray-2")
-              span {{ $t('NN0010') }}
-            div(class="brand-kit-tab-logo__item__menu__row pointer"
-              @click="handleDeleteLogo(logo)")
-              svg-icon(iconName="trash"
-                      iconWidth="24px"
-                      iconColor="gray-2")
-              span {{ $t('NN0034') }}
+image-list(:images="renderedLogos"
+  :showMore="true"
+  @addImage="handleUploadLogo"
+  @handleDownload="handleDownload"
+  @handleDeleteLogo="handleDeleteLogo"
+  @handleOpenMenu="handleOpenMenu"
+  @handleCloseMenu="handleCloseMenu"
+  @loadMore="handleLoadMore")
 </template>
 
 <script lang="ts">
-import ObserverSentinel from '@/components/ObserverSentinel.vue'
+import ImageList, { spItem } from '@/components/image-gallery/ImageList.vue'
 import { IBrand, IBrandLogo } from '@/interfaces/brandkit'
 import brandkitUtils from '@/utils/brandkitUtils'
 import uploadUtils from '@/utils/uploadUtils'
 import vClickOutside from 'click-outside-vue3'
 import { defineComponent } from 'vue'
 import { mapActions, mapGetters } from 'vuex'
+
+interface IRenderedLogos {
+  type: '' | 'add' | 'loading' | 'sentinel'
+  key: string
+  label: string
+  src: string
+  uploading: boolean
+  menuopen: boolean
+  logo?: IBrandLogo
+}
 
 export default defineComponent({
   emits: ['deleteItem'],
@@ -74,7 +39,7 @@ export default defineComponent({
     brandkitUtils.fetchLogos(this.fetchLogos)
   },
   components: {
-    ObserverSentinel
+    ImageList,
   },
   directives: {
     clickOutside: vClickOutside.directive
@@ -96,12 +61,20 @@ export default defineComponent({
     logos(): IBrandLogo[] {
       return (this.currentBrand as IBrand).logos
     },
-    renderedLogos(): (IBrandLogo | string)[] {
-      const res = ['add', ...this.logos]
+    renderedLogos(): IRenderedLogos[] {
+      const res = [spItem('add'), ...this.logos.map(logo => ({
+        type: '' as const,
+        key: logo.id,
+        label: logo.name,
+        src: brandkitUtils.getLogoUrl(logo, this.currentBrand.id, 'tiny'),
+        uploading: logo.id.startsWith('new_'),
+        menuopen: this.menuOpenLogoId === logo.id,
+        logo,
+      }))]
       if (this.isLogosLoading) {
-        res.push('loading')
+        res.push(spItem('loading'))
       } else if (this.logosPageIndex >= 0) {
-        res.push('sentinel')
+        res.push(spItem('sentinel'))
       }
       return res
     }
@@ -112,26 +85,25 @@ export default defineComponent({
       fetchMoreLogos: 'fetchMoreLogos',
       refreshLogoAsset: 'refreshLogoAsset'
     }),
-    getUrl(logo: IBrandLogo): string {
-      return brandkitUtils.getLogoUrl(logo, this.currentBrand.id, 'tiny')
-    },
     checkMenuOpen(logo: IBrandLogo): boolean {
       return this.menuOpenLogoId === logo.id
     },
-    checkUploading(logo: IBrandLogo) {
-      return logo.id.startsWith('new_')
-    },
-    handleOpenMenu(logo: IBrandLogo) {
+    handleOpenMenu({ logo }: IRenderedLogos) {
+      if (!logo) return
       if (this.checkMenuOpen(logo)) {
         this.menuOpenLogoId = ''
       } else {
         this.menuOpenLogoId = logo.id
       }
     },
+    handleCloseMenu() {
+      this.menuOpenLogoId = ''
+    },
     handleUploadLogo() {
       uploadUtils.chooseAssets('logo')
     },
-    handleDownload(logo: IBrandLogo) {
+    handleDownload({ logo }: IRenderedLogos) {
+      if (!logo) return
       const brand = this.currentBrand
       const logoName = logo.name
       const url = brandkitUtils.getLogoUrl(logo, brand.id, 'original')
@@ -151,7 +123,8 @@ export default defineComponent({
         this.startDownloading(url, logoName)
       }
     },
-    handleDeleteLogo(logo: IBrandLogo) {
+    handleDeleteLogo({ logo }: IRenderedLogos) {
+      if (!logo) return
       this.menuOpenLogoId = ''
       this.$emit('deleteItem', {
         type: 'logo',
@@ -172,157 +145,4 @@ export default defineComponent({
 </script>
 
 <style lang="scss" scoped>
-.brand-kit-tab-logo-loading {
-  display: flex;
-  justify-content: center;
-}
-.brand-kit-tab-logo {
-  display: flex;
-  gap: 15px;
-  flex-wrap: wrap;
-  &__item {
-    height: 100px;
-    min-width: 60px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 1px solid setColor(gray-4);
-    box-sizing: border-box;
-    border-radius: 4px;
-    &.add {
-      // width: 100px;
-      display: flex;
-      flex-direction: column;
-      padding: 25px 30px;
-      border: 1px dashed setColor(gray-4);
-      & > .hover {
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        display: none;
-        z-index: 1;
-      }
-      &:hover {
-        background-color: setColor(blue-4);
-        border: 1px solid setColor(blue-4);
-        & > span.primary,
-        & > span.secondary {
-          color: setColor(blue-4);
-        }
-        & > .hover {
-          display: block;
-        }
-      }
-      & > span.primary {
-        @include body-SM;
-        color: setColor(gray-2);
-      }
-      & > span.secondary {
-        @include body-XS;
-        color: setColor(gray-3);
-      }
-    }
-    &__img {
-      height: 100%;
-      width: auto;
-    }
-    &:not(.add):hover,
-    &.hovered {
-      background-color: rgba(setColor(gray-4), 0.5);
-      border: 1px solid setColor(gray-4);
-      & > img {
-        opacity: 0.5;
-      }
-      & > div {
-        display: flex;
-      }
-    }
-    &__more {
-      position: absolute;
-      top: 4px;
-      right: 5px;
-      display: none;
-      align-items: center;
-      justify-content: center;
-      background: white;
-      border-radius: 2px;
-      &-container {
-        width: 24px;
-        height: 24px;
-      }
-    }
-    &__menu {
-      display: flex;
-      flex-direction: column;
-      gap: 10px;
-      padding: 8px 0px;
-      top: calc(100% + 10px);
-      left: 0;
-      width: 216px;
-      position: absolute;
-      background: white;
-      box-shadow: 0px 4px 4px rgba(151, 150, 150, 0.25);
-      border-radius: 5px;
-      z-index: 1;
-      cursor: initial;
-      &__name {
-        height: 25px;
-        padding: 0px 8px;
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-        & > span {
-          @include caption-LG;
-          text-align: left;
-          display: block;
-        }
-      }
-      &__hr {
-        margin: auto;
-        height: 1px;
-        width: calc(100% - 16px);
-        background-color: setColor(gray-4);
-      }
-      &__row {
-        display: flex;
-        align-items: center;
-        gap: 4px;
-        padding: 4px 0px 4px 10px;
-        &:hover {
-          background-color: setColor(blue-4);
-        }
-        & > span {
-          @include body-SM;
-          line-height: 25px;
-          height: 25px;
-          display: block;
-          color: setColor(gray-1);
-        }
-      }
-    }
-  }
-}
-
-.logo-list {
-  &-enter-active,
-  &-leave-active,
-  &-move {
-    &:not(.no-trans) {
-      transition: 0.3s ease;
-    }
-  }
-
-  &-leave-active {
-    position: absolute;
-  }
-
-  &-enter-from,
-  &-leave-to {
-    &:not(.no-trans) {
-      transform: translateY(-30%);
-      opacity: 0;
-    }
-  }
-}
 </style>
